CubeCart TestNG Framework codes part3
