+++ MyPracticeProject/config.properties
@@ -3,7 +3,7 @@ password=automation123!
 produrl=http://cubecartuat.unitedcoder.com/admin_tu8sml.php
 qaurl=http://cubecartuat.unitedcoder.com/admin_tu8sml.php
 reportfolder=doc
-timeout=5
+timeout=10
 username=testautomation1
 qa=1
 screenshot=imageFolder
