--- MyPracticeProject/src/test/java/regressionsuit/testngframework/LoginTestDemo.java
@@ -1,34 +1,0 @@
-package regressionsuit.testngframework;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.PageLoadStrategy;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.Parameters;
-import org.testng.annotations.Test;
-import regressionsuit.pageobjectmodel.DashboardPage;
-import regressionsuit.pageobjectmodel.LoginPage;
-
-public class LoginTestDemo {
-    WebDriver driver;
-
-    @Parameters({"url","username","password"})
-    @Test
-    public void loginTest(String url, String userName,String password){
-            WebDriverManager.chromedriver().setup();
-            driver=new ChromeDriver();
-            driver.navigate().to(url);
-            driver.manage().window().maximize();
-            LoginPage loginPage=new LoginPage(driver);
-            loginPage.Login(userName,password);
-    }
-    @AfterClass
-    public void closeBrowser(){
-        DashboardPage dashboardPage=new DashboardPage(driver);
-        dashboardPage.clickLogOut();
-        driver.close();
-        driver.quit();
-    }
-}
